requirements.txt
flask
flask_restful
markdown
pytest
requests
TinyDB
gunicorn
pika
tensorflow
Pillow
matplotlib
argparse